BigDecimal type in TrendServiceImpl

+++ kanban-backend-app/src/main/java/com/domko/kanbanbackendapp/model/TrendSeries.java
@@ -2,6 +2,7 @@ package com.domko.kanbanbackendapp.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -9,14 +10,14 @@ import java.util.List;
 public class TrendSeries implements Series {
     private String name;
     private String type;
-    private List<Float> data = new ArrayList<>();
+    private List<BigDecimal> data = new ArrayList<>();
 
     public TrendSeries(String name, String type) {
         this.name=name;
         this.type=type;
     }
 
-    public void addAll(List<Float> values){
+    public void addAll(List<BigDecimal> values){
         data.addAll(values);
     }
